getVersionToInstall -> getJLinkState
The return value of getVersionToInstall was unclear to me. E.g. what
would if `installed` was false what would be the meaning of `outdated`
being true or false, since there is nothing installed which can be
outdated. And it what cases can I expect the `versionToBeInstalled` and
`installedVersion` to be defined?

With the four named, distinct return states I tried to clarify this.