Switch to npm trusted publishing
See https://docs.npmjs.com/trusted-publishers

Because trusted publishing requires npm CLI version 11.5.1 or later,
this currently also means, that we have to update npm before running
`npm publish`.

The provenance switch is not needed anymore, as it is automatically
enabled for trusted publishers.

With this, a stored npm token is not needed anymore.

Additionally `npm-release.yml` also needs to be registered as the
trusted publisher in the npm settings for the
@nordicsemiconductor/nrf-jlink-js package.